[Bronze II] Title: 벌집, Time: 100 ms, Memory: 14172 KB -BaekjoonHub

+++ 백준/Bronze/2292. 벌집/벌집.java
@@ -6,7 +6,7 @@ class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
 
-        Long N = Long.parseLong(bufferedReader.readLine());
+        int N = Integer.parseInt(bufferedReader.readLine());
 
         int range = 1;
         int count = 1;
